fix: graphQl `removeCart` mutation
