+++ client-app/core/composables/usePageHead.ts
@@ -1,9 +1,18 @@
 import { eagerComputed, MaybeRef } from "@vueuse/core";
 import { HeadAttrs, useHead } from "@vueuse/head";
 import { unref } from "vue";
-import { UsePageSeoData, useThemeContext } from "@/core";
+import { useThemeContext } from "@/core";
 
-export default function usePageHead(data: UsePageSeoData) {
+export interface IUsePageSeoData {
+  /**
+   * input chunks: ["title_part_1", "title_part_2"]
+   * output string: title_part_1<page_title_divider>title_part_2
+   */
+  title?: MaybeRef<string | string[] | undefined>;
+  meta?: Record<string, MaybeRef<string | undefined>>;
+}
+
+export function usePageHead(data: IUsePageSeoData) {
   const { themeContext } = useThemeContext();
 
   const {
