+++ client-app/xapi/graphql/cart/mutations/removeCart/index.ts
@@ -1,7 +1,9 @@
+import globals from "@/core/globals";
 import { Mutations, MutationsRemoveCartArgs } from "@/xapi/types";
 import mutationDocument from "./removeCartMutation.graphql";
 
 export default async function removeCart(cartId: string): Promise<void> {
+  const { userId } = globals;
   const { $graphqlClient } = useNuxtApp();
 
   await $graphqlClient.mutate<Required<Pick<Mutations, "removeCart">>, MutationsRemoveCartArgs>({
@@ -9,6 +11,7 @@ export default async function removeCart(cartId: string): Promise<void> {
     variables: {
       command: {
         cartId,
+        userId,
       },
     },
   });
