+++ package.json
@@ -84,6 +84,7 @@
     "gulp-zip": "^5.1.0",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.2",
+    "optionator": "^0.9.1",
     "postcss": "^8.4.21",
     "prettier": "^2.8.4",
     "sass": "^1.58.3",
